requirements.txt
python-jobspy
jobsparser
openai
python-dotenv
